Move everything todo /w lego acme to route_manager
This is so all the old lego code is in one place and we can start
reworking it
